Feat/seo page metadata (#217)
* feat(frontend): add SEO metadata to all pages

Split each page into server + client components to enable Next.js metadata
exports. Static pages (home, about, statistics) get dedicated title /
description / keywords. Film and festival pages now generate dynamic
metadata server-side from the API.

- Add frontend/src/lib/seo.ts with site constants and metadata helpers
- Convert each page.tsx to a server component delegating to a *Client
  component (HomeClient, AboutClient, StatisticsClient, FilmClient,
  FestivalClient)
- Implement generateMetadata for films/[slug] and festivals/[slug]
- Add INTERNAL_API_URL so SSR fetches can reach the backend container
  through the Docker network (localhost would resolve to the frontend
  container itself)


* feat(frontend): expand SEO keywords with crew roles and parity indicators

Distribute the additional keyword list across pages:
- DEFAULT_KEYWORDS: add genre, diversité, inclusion, représentation
- STATISTICS_KEYWORDS: add all crew role terms (cheffe de poste,
  réalisatrice, compositrice, costumière, décoratrice, directrice photo,
  directrice production, directrice casting, effets spéciaux, ingénieure
  son, monteuse, scénariste, productrice) and indicators (bonus parité,
  budget moyen)


* chore(frontend): apply Next.js 16 tsconfig changes

Auto-applied by Next.js 16 on dev/build:
- jsx: "preserve" -> "react-jsx" (modern JSX runtime)
- include: add dist/dev/types/**/*.ts


---------
