+++ docker-compose.yaml
@@ -52,6 +52,7 @@ services:
     environment:
       - ENV
       - NEXT_PUBLIC_API_URL
+      - INTERNAL_API_URL=http://backend:5001
     depends_on:
       - backend
